+++ net-spider-rpl/ChangeLog.md
@@ -10,6 +10,7 @@
     * `MergedDIOLink`
     * `DAONode`
     * `DAOLink`
+* Add `ToJSON` instance to `IPv6ID`.
 
 ## 0.4.0.1  -- 2019-12-29