+++ net-spider-rpl/ChangeLog.md
@@ -10,6 +10,8 @@
     * `MergedDIOLink`
     * `DAONode`
     * `DAOLink`
+    * `CombinedNode`
+    * `CombinedLink`
 * Add `ToJSON` instance to `IPv6ID`.
 
 ## 0.4.0.1  -- 2019-12-29